+++ src/main/java/com/example/SmartBuildingBackend/dto/RoomDto.java
@@ -16,7 +16,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class RoomDto {
-    private Integer roomId;
+    private Long roomId;
 
     @NotBlank(message = "Room name is required")
     private String roomName;
