+++ src/main/java/com/example/SmartBuildingBackend/repository/LogAqaraRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.SmartBuildingBackend.entity.LogAqara;
 
-public interface LogAqaraRepository extends JpaRepository<LogAqara, Integer> {
+public interface LogAqaraRepository extends JpaRepository<LogAqara, Long> {
 
 }
