+++ src/main/java/com/example/SmartBuildingBackend/service/FloorService.java
@@ -6,8 +6,8 @@ import com.example.SmartBuildingBackend.dto.FloorDto;
 
 public interface FloorService {
     List<FloorDto> getAllFloors();
-    FloorDto getFloorById(int floorId);
-    FloorDto updateFloor(int floorId, FloorDto updateFloor);
-    void deleteFloor(int floorId);
-    FloorDto addFloor(int blockId,FloorDto floorDto);
+    FloorDto getFloorById(Long floorId);
+    FloorDto updateFloor(Long floorId, FloorDto updateFloor);
+    void deleteFloor(Long floorId);
+    FloorDto addFloor(Long blockId,FloorDto floorDto);
 }
